+++ src/components/Picker/Picker.js
@@ -4,6 +4,7 @@ import {StyleSheet, Platform} from 'react-native';
 import {Picker as RNPicker} from '@react-native-picker/picker';
 import LinearGradient from 'react-native-linear-gradient';
 import {useTranslation} from 'react-i18next';
+import {WheelPicker} from '@delightfulstudio/react-native-wheel-picker-android';
 // configs
 import appConfig from 'app-config';
 // helpers
@@ -15,8 +16,9 @@ import {ThemeContext} from 'src/Themes/Theme.context';
 // constants
 import {TypographyType} from 'src/components/base';
 // custom components
-import {Container, TextButton, FlatList} from 'src/components/base';
+import {Container} from 'src/components/base';
 import NoResult from 'src/components/NoResult';
+import AndroidPicker from './AndroidPicker';
 
 const styles = StyleSheet.create({
   pickerMask: {
@@ -36,26 +38,7 @@ const styles = StyleSheet.create({
       },
     }),
   },
-  androidPicker: {
-    flexGrow: 1,
-    paddingVertical: 15,
-  },
-  androidDataRow: {
-    flexDirection: 'row',
-    width: '100%',
-    paddingHorizontal: 15,
-    paddingVertical: 10,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  androidDataRowSelected: {},
-  androidDataText: {
-    textAlign: 'center',
-    // fontWeight: 'bold',
-  },
-  androidDataTextSelected: {
-    fontWeight: 'bold',
-  },
+
   listEmptyContainer: {
     flex: undefined,
     width: undefined,
@@ -65,7 +48,12 @@ const styles = StyleSheet.create({
   listEmptyTitle: {
     paddingHorizontal: 20,
   },
+  wheelPicker: {
+    height: appConfig.device.height * 0.3,
+  },
 });
+
+const ANDROID_PICKER_FONT_SIZE_RATIO = 38 / 14;
 
 const Picker = ({
   refList = undefined,
@@ -77,11 +65,12 @@ const Picker = ({
   androidItemTextStyle = {},
   listEmptyTitleStyle = {},
   onValueChange = (value, index) => {},
-  getAndroidItemLayout = undefined,
   listEmptyTitle = '',
   listEmptyIconName = undefined,
   listEmptyIconBundle = undefined,
   listEmptyIconSize = undefined,
+  androidItemHeight = 50,
+  androidVisibleItemsNumber = undefined,
 }) => {
   const {theme} = useTheme(ThemeContext);
   const {t} = useTranslation();
@@ -92,73 +81,63 @@ const Picker = ({
     };
   }, [theme]);
 
-  const androidSelectedRowStyle = useMemo(() => {
-    return mergeStyles(styles.androidDataRowSelected, {
-      backgroundColor: rgbaToRgb(hexToRgba(theme.color.primary, 0.1)),
-    });
+  const pickerMaskColors = useMemo(() => {
+    return [theme.color.surface, hexToRgba(theme.color.surface, 0)];
   }, [theme]);
 
-  const androidDataTextUnselectedStyle = useMemo(() => {
-    return {
-      color: theme.color.textSecondary,
-    };
-  });
-
-  const androidDataTextSelectedStyle = useMemo(() => {
-    return {
-      color: theme.color.primary,
-    };
-  });
-
-  const languagePickerItemStyle = mergeStyles(styles.languagePickerItem, {
-    color: theme.color.textPrimary,
-  });
-
-  const androidButtonTypoProps = useMemo(() => {
-    return {type: TypographyType.TITLE_SEMI_LARGE};
-  }, []);
+  const pickerItemStyle = useMemo(() => {
+    return {color: theme.color.textPrimary};
+  }, [theme]);
 
   const listEmptyStyle = useMemo(() => {
     return mergeStyles(styles.listEmptyTitle, listEmptyTitleStyle);
   }, [listEmptyTitleStyle]);
 
-  const renderAndroidData = ({item, index}) => {
-    const isSelected = item.value === selectedValue;
+  const androidFormattedData = useMemo(() => {
+    return data.map((item) => item.label);
+  }, [data]);
 
+  const androidInitPosition = useMemo(() => {
+    const position = data.findIndex((item) => {
+      return item.value === selectedValue;
+    });
+    if (position < 0) {
+      return 0;
+    }
+    return position;
+  }, [data, selectedValue]);
+
+  const androidItemTextColor = useMemo(() => {
+    return theme.typography[TypographyType.LABEL_MEDIUM_SECONDARY].color;
+  }, [theme]);
+
+  const androidSelectedItemTextColor = useMemo(() => {
+    const color = theme.color.textPrimary;
+    return color === '#ffffff' ? '#fffffe' : color;
+  }, [theme]);
+
+  const androidItemFontSize = useMemo(() => {
     return (
-      <TextButton
-        key={index}
-        useTouchableHighlight
-        onPress={() => onValueChange(item.value, index)}
-        style={[
-          styles.androidDataRow,
-          isSelected && androidSelectedRowStyle,
-          androidItemStyle,
-        ]}
-        typoProps={androidButtonTypoProps}
-        titleStyle={[
-          styles.androidDataText,
-          androidDataTextUnselectedStyle,
-          isSelected && [
-            styles.androidDataTextSelected,
-            androidDataTextSelectedStyle,
-          ],
-          androidItemTextStyle,
-        ]}>
-        {item.label}
-      </TextButton>
+      theme.typography[TypographyType.LABEL_MEDIUM].fontSize *
+      ANDROID_PICKER_FONT_SIZE_RATIO
     );
+  }, [theme]);
+
+  const onItemAndroidSelected = (itemData) => {
+    const index = itemData.position;
+    const itemValue = data[index]?.value;
+    onValueChange(itemValue, index);
   };
 
   const renderIosItem = () => {
-    return data.map((lang, index) => (
-      <RNPicker.Item key={index} label={lang.label} value={lang.value} />
+    return data.map((item, index) => (
+      <RNPicker.Item key={index} label={item.label} value={item.value} />
     ));
   };
 
   const renderListEmpty = () => {
     return (
-      <Container center>
+      <Container flex center>
         <NoResult
           iconBundle={listEmptyIconBundle}
           iconName={listEmptyIconName}
@@ -171,47 +150,56 @@ const Picker = ({
     );
   };
 
+  if (!data?.length) {
+    return renderListEmpty();
+  }
+
   switch (Platform.OS) {
     case 'ios':
-      return !!data?.length ? (
+      return (
         <Container noBackground style={pickerContainerStyle}>
           <LinearGradient
             style={styles.pickerMask}
-            colors={[theme.color.surface, hexToRgba(theme.color.surface, 0)]}
+            colors={pickerMaskColors}
             locations={[0.3, 1]}
           />
           <RNPicker
             selectedValue={selectedValue}
             style={styles.picker}
-            itemStyle={languagePickerItemStyle}
+            itemStyle={pickerItemStyle}
             onValueChange={onValueChange}>
             {renderIosItem()}
           </RNPicker>
         </Container>
-      ) : (
-        renderListEmpty()
       );
     case 'android':
       return (
-        <Container centerVertical>
-          <FlatList
-            ref={refList}
-            safeLayout
-            style={styles.picker}
-            data={data}
-            scrollEventThrottle={16}
-            showsVerticalScrollIndicator={false}
-            renderItem={renderAndroidData}
-            keyExtractor={(_, index) => index.toString()}
-            ListEmptyComponent={renderListEmpty}
-            onScrollToIndexFailed={(e) => {
-              console.log(e);
-            }}
-            initialNumToRender={androidInitNumToRender}
-            getItemLayout={getAndroidItemLayout}
-            initialScrollIndex={androidInitScrollIndex}
-          />
-        </Container>
+        // <AndroidPicker
+        //   refList={refList}
+        //   data={data}
+        //   itemHeight={androidItemHeight}
+        //   visibleItemsNumber={androidVisibleItemsNumber}
+        //   renderListEmpty={renderListEmpty}
+        //   initNumToRender={androidInitNumToRender}
+        //   initScrollIndex={androidInitScrollIndex}
+        //   itemStyle={androidItemStyle}
+        //   itemTextStyle={androidItemTextStyle}
+        //   selectedItemPosition={androidInitPosition}
+        //   onValueChange={onValueChange}
+        // />
+
+        <WheelPicker
+          style={styles.wheelPicker}
+          data={androidFormattedData}
+          isCurved
+          hideIndicator
+          onItemSelected={onItemAndroidSelected}
+          selectedItemTextColor={androidSelectedItemTextColor}
+          selectedItemPosition={androidInitPosition}
+          itemTextSize={androidItemFontSize}
+          itemTextColor={androidItemTextColor}
+          backgroundColor={theme.color.surface}
+        />
       );
     default:
       return null;